+++ celllists/tests/common.h
@@ -27,11 +27,11 @@
 #define NREP 100
 #define NPOINT 1000
 
-int fill_random_double(unsigned int seed, double* array, size_t size,
+int fill_random_double(unsigned int seed, double* array, int size,
     double low=-0.5, double high=0.5);
-int fill_random_int(unsigned int seed, int* array, size_t size,
+int fill_random_int(unsigned int seed, int* array, int size,
     int begin, int end);
-int fill_random_permutation(unsigned int seed, int* array, size_t size);
+int fill_random_permutation(unsigned int seed, int* array, int size);
 Cell* create_random_cell_nvec(unsigned int seed, int nvec, double scale=1, bool cuboid=false);
 void random_point(unsigned int seed, double* point, double rcut, const double* center,
     double &norm);
